Implement an HTTP route for /path/to/directory/ CLI: support --dir

+++ index.js
@@ -2,13 +2,23 @@ let fs = require('pn/fs')
 let path = require('path')
 let express = require('express')
 let mime = require('mime-types')
+let archiver = require('archiver')
+let mkdir = require('mkdirp')
+let rimraf = require('rimraf')
+let argv = require('yargs').argv
+
+let rootDir = __dirname;
+
+if (argv.dir){
+	rootDir += argv.dir
+}
+
+console.log(`root directory: ${rootDir}`)
 
 const PORT = process.env.PORT || 8000
 
 let app = express()
 app.listen(PORT, () => console.log(`Listening @ http://127.0.0.1:${PORT}`))
-
-let rootDir = __dirname;
 
 let sendHeaders = async (req, res, next) => {
 
@@ -17,6 +27,10 @@ let sendHeaders = async (req, res, next) => {
 	let filePath = req.filePath = path.join(rootDir, req.url)
 
 	console.log(`filePath: ${req.filePath}`)
+
+	req.isDir = !!filePath.match(/\/$/)
+
+	console.log(`is directory: ${req.isDir}`)
 
 	try {
 		req.stat = await fs.stat(filePath);
@@ -37,16 +51,22 @@ let sendHeaders = async (req, res, next) => {
 	}
 
 	if (req.method === 'PUT'){	
-		res.send(405, "file/dir already exist\n")
+		res.status(405).send("file/dir already exist\n")
 		return;
 	}
 
 	let stat = req.stat
 
 	if (stat && stat.isDirectory()){
-		if (req.method === 'GET'){
-			let files = await fs.readdir(filePath)
-			res.json(files);
+		if (req.method === 'GET'
+			|| req.method === 'HEAD'){
+
+			req.files = await fs.readdir(filePath)
+
+			res.setHeader('Content-Type', mime.contentType('application.json'))
+			res.setHeader('Content-Length', stat.size)
+
+			next()
 			return;
 		}	
 	}
@@ -63,8 +83,31 @@ let sendHeaders = async (req, res, next) => {
 
 // GET
 app.get('*', sendHeaders,(req, res) => {
-	let filePath = req.filePath;
-	fs.createReadStream(filePath).pipe(res)
+	try {
+		let filePath = req.filePath
+		if (req.stat.isDirectory()){
+			if (req.headers['accept'] === 'application/x-gtar'){
+				let archive = archiver('zip')
+			    archive.pipe(res);
+			    archive.bulk([{
+		        	expand: true, 
+		        	cwd: filePath, 
+		        	src: ['**'], 
+		        	dest: filePath
+		        }])
+			    archive.finalize()
+			}
+			else {
+				res.json(req.files);
+			}		
+		}
+		else {
+			fs.createReadStream(filePath).pipe(res)
+		}
+	}
+	catch(e){
+		console.log(e.stack)
+	}
 })
 
 // HEAD
@@ -74,19 +117,38 @@ app.head('*', sendHeaders, (req, res, next) => {
 // PUT
 app.put('*', sendHeaders, (req, res) => {
 	let filePath = req.filePath;
-	req.pipe(fs.createWriteStream(filePath))
+
+	if (req.isDir){
+		mkdir(filePath)
+	}
+	else {
+		req.pipe(fs.createWriteStream(filePath))
+	}
+
+	res.end()
 })
 
 // POST
 app.post('*', sendHeaders, (req, res) => {
 	let filePath = req.filePath;
-	fs.truncate(filePath)
-	req.pipe(fs.createWriteStream(filePath))
+	if (req.isDir){
+		res.status(405).send("cannot post a dir\n")
+	}
+	else {
+		fs.truncate(filePath)
+		req.pipe(fs.createWriteStream(filePath))
+	}
 })
 
 // DELETE
 app.delete('*', sendHeaders, (req, res) => { 
 	let filePath = req.filePath;
-	fs.unlink(filePath)
+	if (req.isDir){
+		console.log('deleting ' + filePath)
+		rimraf(filePath, {}, () => console.log(`deleted directory: ${filePath}`));
+	}
+	else {
+		fs.unlink(filePath)
+	}
 	res.end()
 })
